+++ R/IDateTime.R
@@ -59,6 +59,16 @@ unique.IDate =
     as.IDate(NextMethod())
   }
 
+# define this [<- method to prevent base R's internal rbind coercing integer IDate to double, #2008
+`[<-.IDate` = function(x, i, value) {
+  if (!length(value)) return(x)
+  value = as.integer(as.IDate(value))
+  setattr(x, 'class', NULL)
+  x[i] = value
+  setattr(x, 'class', c('IDate', 'Date'))
+  x
+}
+
 # fix for #1315
 as.list.IDate = function(x, ...) NextMethod()